+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Enable negative `lengthAngle` to configure clockwise and counterclockwise charts
 - Add `rx` and `ry` props to set custom chart center coordinates
 - Add `radius` prop
+- Re-evaluate segments size when `paddingAngle` is > 0
 - Make counterclockwise charts by default
 - Make <ReactMinimalPieChart> a pure component